Refresh stocks example
